requirements.txt
anyio==3.5.0
asgi-babel==0.7.0
asgi-tools==0.63.3
asgiref==3.5.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.72.0
frozendict==2.2.0
googletrans==3.1.0a0
grpcio==1.43.0
grpcio-tools==1.43.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
http-router==2.6.5
httpcore==0.9.1
httptools==0.2.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==6.0.1
orjson==3.6.6
parse-accept-language==0.1.2
pip-autoremove==0.10.0
protobuf==3.19.3
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
sentry-sdk==1.5.3
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
